Feature/login

+++ app/src/main/java/dk/eboks/app/domain/models/login/User.kt
@@ -3,7 +3,7 @@ package dk.eboks.app.domain.models.login
 import java.io.Serializable
 
 data class User(
-        var id: Long = -1,
+        var id: Int = -1,
         var name: String = "",
         var emails: ArrayList<ContactPoint> = arrayListOf(ContactPoint()),
         var mobileNumber: ContactPoint? = null,
